refactor: moved data loading code to app package.
